chore: set base-branch for conform checks
Currently, if you post multiple commits in one pull request, the Conform
tool only checks the latest commit. By adding the `--base-branch`
argument, it will check all commits on top of `main`.

Unfortunately, the built-in conform action does not allow overriding the
default arguments, so we have to run the container directly.

Signed-off-by: John Moon <john.moon@voyagertechnologies.com>