refactor: switch from binary test data to JSON
It's bad practive to store binary artifacts in Git. For clarity,
convert the raw VRT files in this repo to JSON5 (to allow commenting).

A few different sections need to be refactored or gated behind the serde
feature being enabled.

To get back to the original raw VRT files (as needed for the CXX demo),
a new example program is provided to convert from JSON to raw VRT.

Signed-off-by: John Moon <john.moon@voyagertechnologies.com>