chore: add deku-log feature for debugging
Currently, it's hard to debug issues that arise when parsing packets.
It's very helpful to enable trace logging in the Deku crate to see
byte-by-byte what's happening during a parse.

To make this easier, re-export the Deku logging feature and add a
section to the README about how to use it.

Signed-off-by: John Moon <john.moon@voyagertechnologies.com>